+++ modules/setStore.js
@@ -1,0 +1,4 @@
+function setStore(books) {
+    localStorage.setItem('booksStore', JSON.stringify(books));
+  }
+export{setStore};
